refactor: 정렬 로직 분리 - useSortableTable custom hook 생성
